requirements.txt
python-dotenv
python_telegram_bot
requests
